Draw the map walls facing the camera
Draw the wall in the player eye direction using simple ray
casting to figure out the distance to the wall. The further the wall,
the smaller it looks.

The camera has a rather small field of view (PI/4) and the player
can't move yet. Ceiling and floor are blank, and wall is drawn using the
pound sign.

There's a lot of code cleanup to do, but it works.