Stop the loop when a key is pressed
Stop the infinite loop when any key is pressed. We'll use that later
on to react to user input, and make the player move and turn around.

This was a tough one, only tested on MacOS. Toggle terminal to
cbreak mode (see https://en.wikipedia.org/wiki/Terminal_mode)
using `stty cbreak` so that we can listen for individual keystrokes
and not wait for a line break. Set the tty back to sane (cooked) mode
at the end of the execution.