Make the player turn around using left and right keys
Listen for keystrokes and make the player turn left or right by adjusting player point of view.

To get proper animation, we need to account for elapsed time between frames. Since we only
get key inputs and not key pressed/released events from the terminal, we also need to limit the
frame rate to match the terminal key input delay.

Finally, we display stats at the top of the screen (showing FPS and player info)
when the SPACE key is pressed.