Draw wall boundaries
Draw wall boundaries that are visible from the player position
to add perspective illusion. To find out which wall edges are
visible, we first compute the normal vector to the wall plane
facing inward, and select those edges for which the dot product
with the view unit vector is positive.